separate testing and update docs
