+++ testing/backend/test_main.py
@@ -454,610 +454,3 @@ def test_repository_project_json_language_stats():
         assert "repo_count" in first_lang
         assert "average_percentage" in first_lang
         assert "total_size" in first_lang
-
-def test_tech_radar_update_no_entries():
-    """Test the tech radar update endpoint with missing entries.
-    
-    This test verifies that the endpoint correctly handles requests with
-    missing entries data by returning an appropriate error response.
-
-    Endpoint:
-        POST /review/api/tech-radar/update
-
-    Expects:
-        - 400 status code
-        - JSON response with error message
-        - Error message indicating invalid or missing title
-    """
-    response = requests.post(f"{BASE_URL}/review/api/tech-radar/update", json={}, timeout=10)
-    assert response.status_code == 400
-    data = response.json()
-    assert "error" in data
-    assert data["error"] == "Invalid or empty entries data"
-
-def test_tech_radar_update_partial():
-    """Test the tech radar update endpoint with a partial update.
-    
-    This test verifies that the endpoint correctly processes updates
-    when provided with the complete tech radar structure.
-
-    Endpoint:
-        POST /review/api/tech-radar/update
-
-    Test Data:
-        - Complete tech radar structure
-        - Valid entries with all required fields
-
-    Expects:
-        - 200 status code
-        - Successful update of entries
-        - Correct structure in stored data
-    """
-    random_number = random.randint(100,1000)
-    test_data = {
-        "entries": [
-            {
-                "id": "test-entry-partial-1",
-                "title": "Test Entry Partial 1",
-                "description": "Languages",
-                "key": "test1",
-                "url": "#",
-                "quadrant": "1",
-                "timeline": [
-                    {
-                        "moved": 0,
-                        "ringId": "ignore",
-                        "date": "2000-01-01",
-                        "description": f"For testing purposes [CASE:{random_number}:1]"
-                    }
-                ],
-                "links": []
-            }
-        ]
-    }
-    
-    response = requests.post(
-        f"{BASE_URL}/review/api/tech-radar/update",
-        json=test_data,
-        timeout=10
-    )
-    assert response.status_code == 200
-    
-    # Verify the updates
-    get_response = requests.get(f"{BASE_URL}/api/tech-radar/json", timeout=10)
-    assert get_response.status_code == 200
-    updated_data = get_response.json()
-    
-    # Verify our entry exists and is correct
-    updated_entries = {entry["id"]: entry for entry in updated_data["entries"]}
-    assert "test-entry-partial-1" in updated_entries
-    assert updated_entries["test-entry-partial-1"]["timeline"][0]["ringId"] == "ignore"
-    assert updated_entries["test-entry-partial-1"]["quadrant"] == "1"
-
-def test_tech_radar_update_invalid_entries():
-    """Test the tech radar update endpoint with invalid entries format.
-    
-    This test verifies that the endpoint correctly handles requests with
-    invalid entries data format by returning an appropriate error response.
-
-    Endpoint:
-        POST /review/api/tech-radar/update
-
-    Test Data:
-        - Invalid entries format
-        - Missing required fields
-        - Malformed data structures
-
-    Expects:
-        - 400 status code
-        - Error message for invalid data
-        - No changes to existing entries
-    """
-    # Test with missing title
-    response = requests.post(
-        f"{BASE_URL}/review/api/tech-radar/update",
-        json={"entries": "not_an_array"},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert response.json()["error"] == "Invalid or empty entries data"
-
-def test_tech_radar_update_invalid_structure():
-    """Test the tech radar update endpoint with invalid structure.
-    
-    This test verifies that the endpoint correctly validates the complete
-    structure of the tech radar data, including title, quadrants, rings,
-    and entries.
-
-    Endpoint:
-        POST /review/api/tech-radar/update
-
-    Test Data:
-        - Missing title
-        - Invalid quadrants structure
-        - Invalid rings structure
-        - Invalid entries structure
-
-    Expects:
-        - 400 status code for each invalid case
-        - Appropriate error messages
-        - No changes to existing data
-    """
-    # Test missing title
-    response = requests.post(
-        f"{BASE_URL}/review/api/tech-radar/update",
-        json={
-            "quadrants": [],
-            "rings": [],
-            "entries": []
-        },
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert response.json()["error"] == "Invalid or empty entries data"
-
-    # Test invalid quadrants
-    response = requests.post(
-        f"{BASE_URL}/review/api/tech-radar/update",
-        json={
-            "title": "Test Radar",
-            "quadrants": [{"invalid": "structure"}],
-            "rings": [],
-            "entries": []
-        },
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert response.json()["error"] == "Invalid or empty entries data"
-
-    # Test invalid rings
-    response = requests.post(
-        f"{BASE_URL}/review/api/tech-radar/update",
-        json={
-            "title": "Test Radar",
-            "quadrants": [{"id": "1", "name": "Test"}],
-            "rings": [{"invalid": "structure"}],
-            "entries": []
-        },
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert response.json()["error"] == "Invalid or empty entries data"
-
-def test_tech_radar_update_valid_structure():
-    """Test the tech radar update endpoint with valid complete structure.
-    
-    This test verifies that the endpoint correctly processes a complete
-    tech radar update with valid structure for all components.
-
-    Endpoint:
-        POST /review/api/tech-radar/update
-
-    Test Data:
-        - Valid title
-        - Valid quadrants with required fields
-        - Valid rings with required fields
-        - Valid entries with required fields
-
-    Expects:
-        - 200 status code
-        - Successful update confirmation
-        - Correct structure in stored data
-    """
-    random_number = random.randint(100,1000)
-    test_data = {
-        "entries": [
-            {
-                "id": "test-entry-1",
-                "title": "Test Entry 1",
-                "description": "Languages",
-                "key": "test1",
-                "url": "#",
-                "quadrant": "1",
-                "timeline": [
-                    {
-                        "moved": 0,
-                        "ringId": "ignore",
-                        "date": "2000-01-01",
-                        "description": f"For testing purposes [CASE:{random_number}:2]"
-                    }
-                ],
-                "links": []
-            }
-        ]
-    }
-    
-    response = requests.post(
-        f"{BASE_URL}/review/api/tech-radar/update",
-        json=test_data,
-        timeout=10
-    )
-    assert response.status_code == 200
-    assert response.json()["message"] == "Tech radar updated successfully"
-
-    # Verify the update
-    get_response = requests.get(f"{BASE_URL}/api/tech-radar/json", timeout=10)
-    assert get_response.status_code == 200
-    updated_data = get_response.json()
-
-    # Verify entry structure
-    entries = updated_data["entries"]
-    test_entry = next((entry for entry in entries if entry["id"] == "test-entry-1"), None)
-    assert test_entry is not None, "No entry with id 'test-entry-1' found"
-    assert str(random_number) in test_entry["timeline"][0]["description"], "Entry with id 'test-entry-1' does not have the expected description"
-
-def test_tech_radar_update_invalid_references():
-    """Test the tech radar update endpoint with invalid references.
-    
-    This test verifies that the endpoint correctly validates references
-    between entries and their quadrants/rings.
-
-    Endpoint:
-        POST /review/api/tech-radar/update
-
-    Test Data:
-        - Entry with invalid quadrant reference
-        - Entry with invalid ring reference
-        - Entry with missing required fields
-
-    Expects:
-        - 400 status code
-        - Appropriate error messages
-        - No changes to existing data
-    """
-    test_data = {
-        "title": "ONS Tech Radar",
-        "quadrants": [
-            {"id": "1", "name": "Languages"}
-        ],
-        "rings": [
-            {"id": "adopt", "name": "ADOPT", "color": "#008a00"}
-        ],
-        "entries": [
-            {
-                "id": "test-entry",
-                "title": "Test Entry",
-                "quadrant": "invalid",  # Invalid quadrant reference
-                "timeline": [
-                    {
-                        "moved": 0,
-                        "ringId": "invalid",  # Invalid ring reference
-                        "date": "2024-03",
-                        "description": "Test"
-                    }
-                ]
-            }
-        ]
-    }
-    
-    response = requests.post(
-        f"{BASE_URL}/review/api/tech-radar/update",
-        json=test_data,
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert "Invalid entry structure" in response.json()["error"]
-
-def test_admin_banner_get():
-    """Test the admin banners endpoint for retrieving banner messages.
-    
-    This test verifies that the endpoint correctly returns banner messages
-    from the S3 bucket. It checks the structure of the response and ensures
-    the messages array is present.
-
-    Endpoint:
-        GET /admin/api/banners
-
-    Expects:
-        - 200 status code
-        - JSON response containing a messages array
-        - Valid structure that can be parsed by the admin UI
-    """
-    response = requests.get(f"{BASE_URL}/admin/api/banners", timeout=10)
-    assert response.status_code == 200
-    data = response.json()
-    
-    # Verify the response structure
-    assert "messages" in data
-    assert isinstance(data["messages"], list)
-
-def test_admin_banner_update():
-    """Test the admin banners update endpoint.
-    
-    This test verifies that the endpoint correctly processes banner 
-    updates with valid data and saves them to the S3 bucket.
-
-    Endpoint:
-        POST /admin/api/banners/update
-
-    Test Data:
-        - Valid banner message
-        - Array of pages where the banner should appear
-        - Optional banner title and type
-
-    Expects:
-        - 200 status code
-        - Success message confirming the banner was added
-        - Banner should be retrievable in subsequent GET requests
-    """
-    test_data = {
-        "banner": {
-            "message": "Test Banner Message",
-            "title": "Test Banner",
-            "type": "info",
-            "pages": ["home", "tech-radar"],
-            "show": True
-        }
-    }
-    
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/update",
-        json=test_data,
-        timeout=10
-    )
-    assert response.status_code == 200
-    data = response.json()
-    assert "message" in data
-    assert data["message"] == "Banner added successfully"
-    
-    # Verify the banner was added
-    get_response = requests.get(f"{BASE_URL}/admin/api/banners", timeout=10)
-    assert get_response.status_code == 200
-    get_data = get_response.json()
-    assert "messages" in get_data
-    
-    # Find the added banner by message
-    added_banner = next((banner for banner in get_data["messages"] 
-                         if banner["message"] == "Test Banner Message"), None)
-    assert added_banner is not None
-    assert added_banner["title"] == "Test Banner"
-    assert added_banner["type"] == "info"
-    assert "home" in added_banner["pages"]
-    assert "tech-radar" in added_banner["pages"]
-    assert added_banner["show"] is True
-
-def test_admin_banner_update_invalid():
-    """Test the admin banners update endpoint with invalid data.
-    
-    This test verifies that the endpoint correctly validates banner data
-    and returns appropriate error responses for invalid inputs.
-
-    Endpoint:
-        POST /admin/api/banners/update
-
-    Test Data:
-        - Missing message field
-        - Empty pages array
-        - Malformed banner object
-
-    Expects:
-        - 400 status code
-        - Error message indicating invalid banner data
-    """
-    # Test with missing message
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/update",
-        json={"banner": {"pages": ["home"]}},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert response.json()["error"] == "Invalid banner data"
-    
-    # Test with empty pages array
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/update",
-        json={"banner": {"message": "Test", "pages": []}},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert response.json()["error"] == "Invalid banner data"
-    
-    # Test with malformed request body
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/update",
-        json={"not_banner": {}},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert response.json()["error"] == "Invalid banner data"
-
-def test_admin_banner_toggle():
-    """Test the admin banners toggle endpoint.
-    
-    This test verifies that the endpoint correctly toggles banner visibility
-    by updating the 'show' property in the S3 bucket.
-
-    Endpoint:
-        POST /admin/api/banners/toggle
-
-    Test Data:
-        - Valid index of an existing banner
-        - New visibility state
-
-    Expects:
-        - 200 status code
-        - Success message confirming the visibility update
-        - Banner visibility should be updated in subsequent GET requests
-    """
-    # First, add a test banner
-    test_data = {
-        "banner": {
-            "message": "Toggle Test Banner",
-            "pages": ["home"],
-            "show": True
-        }
-    }
-    
-    # Add the banner
-    requests.post(
-        f"{BASE_URL}/admin/api/banners/update",
-        json=test_data,
-        timeout=10
-    )
-    
-    # Get all banners
-    get_response = requests.get(f"{BASE_URL}/admin/api/banners", timeout=10)
-    assert get_response.status_code == 200
-    banners = get_response.json()["messages"]
-    
-    # Find the index of our test banner
-    test_banner_index = next((i for i, banner in enumerate(banners) 
-                             if banner["message"] == "Toggle Test Banner"), None)
-    assert test_banner_index is not None
-    
-    # Toggle the banner visibility
-    toggle_data = {
-        "index": test_banner_index,
-        "show": False
-    }
-    
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/toggle",
-        json=toggle_data,
-        timeout=10
-    )
-    assert response.status_code == 200
-    data = response.json()
-    assert data["message"] == "Banner visibility updated successfully"
-    
-    # Verify the banner was toggled
-    get_response = requests.get(f"{BASE_URL}/admin/api/banners", timeout=10)
-    updated_banners = get_response.json()["messages"]
-    assert updated_banners[test_banner_index]["show"] is False
-
-def test_admin_banner_toggle_invalid():
-    """Test the admin banners toggle endpoint with invalid data.
-    
-    This test verifies that the endpoint correctly validates toggle data
-    and returns appropriate error responses for invalid inputs.
-
-    Endpoint:
-        POST /admin/api/banners/toggle
-
-    Test Data:
-        - Invalid index (non-numeric)
-        - Out of range index
-        - Missing index
-
-    Expects:
-        - 400 status code
-        - Error message indicating invalid index
-    """
-    # Test with string index
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/toggle",
-        json={"index": "not-a-number", "show": True},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert "Invalid banner index" in response.json()["error"]
-    
-    # Test with missing index
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/toggle",
-        json={"show": True},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert "Invalid banner index" in response.json()["error"]
-
-def test_admin_banner_delete():
-    """Test the admin banners delete endpoint.
-    
-    This test verifies that the endpoint correctly deletes banners
-    from the S3 bucket based on their index.
-
-    Endpoint:
-        POST /admin/api/banners/delete
-
-    Test Data:
-        - Valid index of an existing banner
-
-    Expects:
-        - 200 status code
-        - Success message confirming deletion
-        - Banner should be removed in subsequent GET requests
-    """
-    # First, add a test banner to delete
-    test_data = {
-        "banner": {
-            "message": "Delete Test Banner",
-            "pages": ["home"]
-        }
-    }
-    
-    # Add the banner
-    requests.post(
-        f"{BASE_URL}/admin/api/banners/update",
-        json=test_data,
-        timeout=10
-    )
-    
-    # Get all banners
-    get_response = requests.get(f"{BASE_URL}/admin/api/banners", timeout=10)
-    assert get_response.status_code == 200
-    banners = get_response.json()["messages"]
-    
-    # Find the index of our test banner
-    test_banner_index = next((i for i, banner in enumerate(banners) 
-                             if banner["message"] == "Delete Test Banner"), None)
-    assert test_banner_index is not None
-    
-    # Delete the banner
-    delete_data = {
-        "index": test_banner_index
-    }
-    
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/delete",
-        json=delete_data,
-        timeout=10
-    )
-    assert response.status_code == 200
-    data = response.json()
-    assert data["message"] == "Banner deleted successfully"
-    
-    # Verify the banner was deleted
-    get_response = requests.get(f"{BASE_URL}/admin/api/banners", timeout=10)
-    updated_banners = get_response.json()["messages"]
-    
-    # The banner should no longer exist
-    deleted_banner = next((banner for banner in updated_banners 
-                          if banner["message"] == "Delete Test Banner"), None)
-    assert deleted_banner is None
-
-def test_admin_banner_delete_invalid():
-    """Test the admin banners delete endpoint with invalid data.
-    
-    This test verifies that the endpoint correctly validates delete data
-    and returns appropriate error responses for invalid inputs.
-
-    Endpoint:
-        POST /admin/api/banners/delete
-
-    Test Data:
-        - Invalid index (non-numeric)
-        - Out of range index
-        - Missing index
-
-    Expects:
-        - 400 status code
-        - Error message indicating invalid index
-    """
-    # Test with string index
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/delete",
-        json={"index": "not-a-number"},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert "Invalid banner index" in response.json()["error"]
-
-    # Test with missing index
-    response = requests.post(
-        f"{BASE_URL}/admin/api/banners/delete",
-        json={},
-        timeout=10
-    )
-    assert response.status_code == 400
-    assert "Invalid banner index" in response.json()["error"]
