+++ testing/pytest.ini
@@ -1,0 +1,9 @@
+[pytest]
+testpaths = backend
+python_files = test_*.py
+python_functions = test_*
+filterwarnings =
+    ignore::DeprecationWarning
+    ignore::PendingDeprecationWarning
+log_cli = true
+log_cli_level = INFO
